Move jQuery plugins into modules and add them to the bundle
